added 'list' dir. in common; moved the files into it <(|_|)>
