+++ arakelyan.daniil/S1/inputData.hpp
@@ -2,7 +2,7 @@
 #define INPUTDATA_HPP
 
 #include <iostream>
-#include <binList.hpp>
+#include <list/binList.hpp>
 
 namespace arakelyan
 {
