+++ arakelyan.daniil/S1/outputData.hpp
@@ -2,7 +2,7 @@
 #define OUTPUTPROC_HPP
 
 #include <iostream>
-#include <binList.hpp>
+#include <list/binList.hpp>
 
 namespace arakelyan
 {
